# perf: performance improvements