--- a/packages/logger/src/utils/colorize.ts
+++ b/packages/logger/src/utils/colorize.ts
@@ -1,6 +1,17 @@
 import { Color, OgmaSimpleType } from '@ogma/common';
 import { style as styler, Styler } from '@ogma/styler';
 
+const colorizeMap: Record<Color, string> = {
+  [Color.BLACK]: 'black',
+  [Color.RED]: 'red',
+  [Color.GREEN]: 'green',
+  [Color.YELLOW]: 'yellow',
+  [Color.BLUE]: 'blue',
+  [Color.MAGENTA]: 'magenta',
+  [Color.CYAN]: 'cyan',
+  [Color.WHITE]: 'white',
+};
+
 export function colorize(
   value: OgmaSimpleType,
   color: Color = Color.WHITE,
@@ -8,7 +19,8 @@ export function colorize(
   useColor = true,
 ): string {
   if (useColor) {
-    value = style[Color[color].toLowerCase()].apply(value);
+    return style[colorizeMap[color]].apply(value);
   }
+
   return value.toString();
 }