--- a/packages/styler/src/styler.ts
+++ b/packages/styler/src/styler.ts
@@ -2,151 +2,210 @@ import { OgmaStream } from '@ogma/common';
 
 import { Style } from './style.enum';
 
+const sgrColorMap = new Map([
+  ['3', true],
+  ['4', true],
+  ['9', true],
+]);
+
+const colorMap: Record<Style, string> = {
+  [Style.BLACK]: Style.BLACK.toString(),
+  [Style.BLACKBG]: Style.BLACKBG.toString(),
+  [Style.BRIGHTBLACK]: Style.BRIGHTBLACK.toString(),
+  [Style.BRIGHTBLACKBG]: Style.BRIGHTBLACKBG.toString(),
+  [Style.RED]: Style.RED.toString(),
+  [Style.REDBG]: Style.REDBG.toString(),
+  [Style.BRIGHTRED]: Style.BRIGHTRED.toString(),
+  [Style.BRIGHTREDBG]: Style.BRIGHTREDBG.toString(),
+  [Style.GREEN]: Style.GREEN.toString(),
+  [Style.GREENBG]: Style.GREENBG.toString(),
+  [Style.BRIGHTGREEN]: Style.BRIGHTGREEN.toString(),
+  [Style.BRIGHTGREENBG]: Style.BRIGHTGREENBG.toString(),
+  [Style.YELLOW]: Style.YELLOW.toString(),
+  [Style.YELLOWBG]: Style.YELLOWBG.toString(),
+  [Style.BRIGHTYELLOW]: Style.BRIGHTYELLOW.toString(),
+  [Style.BRIGHTYELLOWBG]: Style.BRIGHTYELLOWBG.toString(),
+  [Style.BLUE]: Style.BLUE.toString(),
+  [Style.BLUEBG]: Style.BLUEBG.toString(),
+  [Style.BRIGHTBLUE]: Style.BRIGHTBLUE.toString(),
+  [Style.MAGENTA]: Style.MAGENTA.toString(),
+  [Style.MAGENTABG]: Style.MAGENTABG.toString(),
+  [Style.BRIGHTMAGENTA]: Style.BRIGHTMAGENTA.toString(),
+  [Style.BRIGHTMAGENTABG]: Style.BRIGHTMAGENTABG.toString(),
+  [Style.CYAN]: Style.CYAN.toString(),
+  [Style.CYANBG]: Style.CYANBG.toString(),
+  [Style.BRIGHTCYAN]: Style.BRIGHTCYAN.toString(),
+  [Style.BIRGHTCYANBG]: Style.BIRGHTCYANBG.toString(),
+  [Style.WHITE]: Style.WHITE.toString(),
+  [Style.WHITEBG]: Style.WHITEBG.toString(),
+  [Style.BRIGHTWHIET]: Style.BRIGHTWHIET.toString(),
+  [Style.BRIGHTWHITEBG]: Style.BRIGHTWHITEBG.toString(),
+  [Style.BOLD]: Style.BOLD.toString(),
+  [Style.FAINT]: Style.FAINT.toString(),
+  [Style.ITALIC]: Style.ITALIC.toString(),
+  [Style.UNDERLINE]: Style.UNDERLINE.toString(),
+  [Style.BLINK]: Style.BLINK.toString(),
+  [Style.FASTBLINK]: Style.FASTBLINK.toString(),
+  [Style.INVERT]: Style.INVERT.toString(),
+  [Style.STRIKE]: Style.STRIKE.toString(),
+  [Style.DOUBLEUNDERLINE]: Style.DOUBLEUNDERLINE.toString(),
+  [Style.REVEAL]: Style.REVEAL.toString(),
+  [Style.FRAMED]: Style.FRAMED.toString(),
+  [Style.ENCIRCLED]: Style.ENCIRCLED.toString(),
+  [Style.OVERLINED]: Style.OVERLINED.toString(),
+  [Style.SUPER]: Style.SUPER.toString(),
+  [Style.SUB]: Style.SUB.toString(),
+  [Style.NONE]: Style.NONE.toString(),
+  [Style.CONCEAL]: Style.CONCEAL.toString(),
+  [Style.COLOR]: Style.COLOR.toString(),
+  [Style.BGCOLOR]: Style.BGCOLOR.toString(),
+  [Style.BRIGHTBLUEBG]: Style.BRIGHTBLUEBG.toString(),
+};
+
 export class Styler {
-  private stylesToApply = [];
+  private stylesToApply = '';
   private colorDepth = 16;
   private useStyle = true;
   public get black() {
-    return this.sgr(Style.BLACK);
+    return this.sgr(colorMap[Style.BLACK]);
   }
   public get blackBg() {
-    return this.sgr(Style.BLACKBG);
+    return this.sgr(colorMap[Style.BLACKBG]);
   }
   public get bBlack() {
-    return this.sgr(Style.BRIGHTBLACK);
+    return this.sgr(colorMap[Style.BRIGHTBLACK]);
   }
   public get bBlackBg() {
-    return this.sgr(Style.BRIGHTBLACKBG);
+    return this.sgr(colorMap[Style.BRIGHTBLACKBG]);
   }
   public get red() {
-    return this.sgr(Style.RED);
+    return this.sgr(colorMap[Style.RED]);
   }
   public get redBg() {
-    return this.sgr(Style.REDBG);
+    return this.sgr(colorMap[Style.REDBG]);
   }
   public get bRed() {
-    return this.sgr(Style.BRIGHTRED);
+    return this.sgr(colorMap[Style.BRIGHTRED]);
   }
   public get bRedBg() {
-    return this.sgr(Style.BRIGHTREDBG);
+    return this.sgr(colorMap[Style.BRIGHTREDBG]);
   }
   public get green() {
-    return this.sgr(Style.GREEN);
+    return this.sgr(colorMap[Style.GREEN]);
   }
   public get greenBg() {
-    return this.sgr(Style.GREENBG);
+    return this.sgr(colorMap[Style.GREENBG]);
   }
   public get bGreen() {
-    return this.sgr(Style.BRIGHTGREEN);
+    return this.sgr(colorMap[Style.BRIGHTGREEN]);
   }
   public get bGreenBg() {
-    return this.sgr(Style.BRIGHTGREENBG);
+    return this.sgr(colorMap[Style.BRIGHTGREENBG]);
   }
   public get yellow() {
-    return this.sgr(Style.YELLOW);
+    return this.sgr(colorMap[Style.YELLOW]);
   }
   public get yellowBg() {
-    return this.sgr(Style.YELLOWBG);
+    return this.sgr(colorMap[Style.YELLOWBG]);
   }
   public get bYellow() {
-    return this.sgr(Style.BRIGHTYELLOW);
+    return this.sgr(colorMap[Style.BRIGHTYELLOW]);
   }
-
   public get bYellowBg() {
-    return this.sgr(Style.BRIGHTYELLOWBG);
+    return this.sgr(colorMap[Style.BRIGHTYELLOWBG]);
   }
   public get blue() {
-    return this.sgr(Style.BLUE);
+    return this.sgr(colorMap[Style.BLUE]);
   }
   public get blueBg() {
-    return this.sgr(Style.BLUEBG);
+    return this.sgr(colorMap[Style.BLUEBG]);
   }
   public get bBlue() {
-    return this.sgr(Style.BRIGHTBLUE);
+    return this.sgr(colorMap[Style.BRIGHTBLUE]);
   }
   public get bBlueBg() {
-    return this.sgr(Style.BRIGHTBLUEBG);
+    return this.sgr(colorMap[Style.BRIGHTBLUEBG]);
   }
   public get magenta() {
-    return this.sgr(Style.MAGENTA);
+    return this.sgr(colorMap[Style.MAGENTA]);
   }
   public get magentaBg() {
-    return this.sgr(Style.MAGENTABG);
+    return this.sgr(colorMap[Style.MAGENTABG]);
   }
   public get bMagenta() {
-    return this.sgr(Style.BRIGHTMAGENTA);
+    return this.sgr(colorMap[Style.BRIGHTMAGENTA]);
   }
   public get bMagentaBg() {
-    return this.sgr(Style.BRIGHTMAGENTABG);
+    return this.sgr(colorMap[Style.BRIGHTMAGENTABG]);
   }
   public get cyan() {
-    return this.sgr(Style.CYAN);
+    return this.sgr(colorMap[Style.CYAN]);
   }
   public get cyanBg() {
-    return this.sgr(Style.CYANBG);
+    return this.sgr(colorMap[Style.CYANBG]);
   }
   public get bCyan() {
-    return this.sgr(Style.BRIGHTCYAN);
+    return this.sgr(colorMap[Style.BRIGHTCYAN]);
   }
   public get bCyanBg() {
-    return this.sgr(Style.BIRGHTCYANBG);
+    return this.sgr(colorMap[Style.BIRGHTCYANBG]);
   }
   public get white() {
-    return this.sgr(Style.WHITE);
+    return this.sgr(colorMap[Style.WHITE]);
   }
   public get whiteBg() {
-    return this.sgr(Style.WHITEBG);
+    return this.sgr(colorMap[Style.WHITEBG]);
   }
   public get bWhite() {
-    return this.sgr(Style.BRIGHTWHIET);
+    return this.sgr(colorMap[Style.BRIGHTWHIET]);
   }
   public get bWhiteBg() {
-    return this.sgr(Style.BRIGHTWHITEBG);
+    return this.sgr(colorMap[Style.BRIGHTWHITEBG]);
   }
   public get bold() {
-    return this.sgr(Style.BOLD);
+    return this.sgr(colorMap[Style.BOLD]);
   }
   public get faint() {
-    return this.sgr(Style.FAINT);
+    return this.sgr(colorMap[Style.FAINT]);
   }
   public get italic() {
-    return this.sgr(Style.ITALIC);
+    return this.sgr(colorMap[Style.ITALIC]);
   }
   public get underline() {
-    return this.sgr(Style.UNDERLINE);
+    return this.sgr(colorMap[Style.UNDERLINE]);
   }
   public get blink() {
-    return this.sgr(Style.BLINK);
+    return this.sgr(colorMap[Style.BLINK]);
   }
   public get fastBlink() {
-    return this.sgr(Style.FASTBLINK);
+    return this.sgr(colorMap[Style.FASTBLINK]);
   }
   public get invert() {
-    return this.sgr(Style.INVERT);
+    return this.sgr(colorMap[Style.INVERT]);
   }
   public get strikeThrough() {
-    return this.sgr(Style.STRIKE);
+    return this.sgr(colorMap[Style.STRIKE]);
   }
   public get doubleUnderline() {
-    return this.sgr(Style.DOUBLEUNDERLINE);
+    return this.sgr(colorMap[Style.DOUBLEUNDERLINE]);
   }
   public get reveal() {
-    return this.sgr(Style.REVEAL);
+    return this.sgr(colorMap[Style.REVEAL]);
   }
   public get framed() {
-    return this.sgr(Style.FRAMED);
+    return this.sgr(colorMap[Style.FRAMED]);
   }
   public get encircled() {
-    return this.sgr(Style.ENCIRCLED);
+    return this.sgr(colorMap[Style.ENCIRCLED]);
   }
   public get overlined() {
-    return this.sgr(Style.OVERLINED);
+    return this.sgr(colorMap[Style.OVERLINED]);
   }
   public get superscript() {
-    return this.sgr(Style.SUPER);
+    return this.sgr(colorMap[Style.SUPER]);
   }
   public get subscript() {
-    return this.sgr(Style.SUB);
+    return this.sgr(colorMap[Style.SUB]);
   }
 
   public constructor(stream: Pick<OgmaStream, 'getColorDepth'> = process.stdout) {
@@ -172,12 +231,13 @@ export class Styler {
    * @returns A string that has the proper SGR values
    */
   public apply(val: string | number | boolean) {
-    let retString = `${this.stylesToApply.join('')}${val}`;
-    if (this.stylesToApply.length) {
-      retString += '\x1B[0m';
-    }
-    this.stylesToApply = [];
-    return retString;
+    if (this.stylesToApply === '') return val.toString();
+
+    const returnValue = `${this.stylesToApply}${val}\x1B[0m`;
+
+    this.stylesToApply = '';
+
+    return returnValue;
   }
 
   /**
@@ -224,18 +284,18 @@ export class Styler {
   /**
    * @link https://en.wikipedia.org/wiki/ANSI_escape_code#SGR
    */
-  private sgr(val: number | string): this {
+  private sgr(val: string): this {
     if (
-      (this.colorDepth === 1 &&
-        val.toString().length >= 2 &&
-        ['3', '4', '9'].includes(val.toString()[0])) ||
-      val.toString().substr(0, 2) === '10'
+      (this.colorDepth === 1 && val.length >= 2 && sgrColorMap.has(val[0])) ||
+      (val[0] === '1' && val[1] === '0')
     ) {
       return this;
     }
+
     if (this.useStyle) {
-      this.stylesToApply.push(`\x1B[${val.toString()}m`);
+      this.stylesToApply += `\x1B[${val}m`;
     }
+
     return this;
   }
 }