Hello!

I did some benchmarks on your library and here what I found, the initial profiler was:

```
Statistical profiling result from ogma-initial-solution.log, (11121 ticks, 1 unaccounted, 0 excluded).

 [Shared libraries]:
   ticks  total  nonlib   name
   2619   23.6%          /home/h4ad/.asdf/installs/nodejs/16.16.0/bin/node
    525    4.7%          /usr/lib/x86_64-linux-gnu/libc-2.31.so
     54    0.5%          [vdso]
      1    0.0%          /usr/lib/x86_64-linux-gnu/libstdc++.so.6.0.28

 [JavaScript]:
   ticks  total  nonlib   name
    130    1.2%    1.6%  LazyCompile: *formatStream /home/h4ad/Projects/opensource/performance-analysis/ogma-performance-analysis/node_modules/@ogma/logger/src/logger/ogma.js:151:17
     70    0.6%    0.9%  LazyCompile: *<anonymous> /home/h4ad/Projects/opensource/performance-analysis/ogma-performance-analysis/dist/ogma-log.stress.js:1:1
     59    0.5%    0.7%  LazyCompile: *_write node:internal/streams/writable:283:16
     47    0.4%    0.6%  LazyCompile: *get magenta /home/h4ad/Projects/opensource/performance-analysis/ogma-performance-analysis/node_modules/@ogma/styler/lib/styler.js:83:16
     41    0.4%    0.5%  LazyCompile: *get cyan /home/h4ad/Projects/opensource/performance-analysis/ogma-performance-analysis/node_modules/@ogma/styler/lib/styler.js:95:13
     41    0.4%    0.5%  LazyCompile: *apply /home/h4ad/Projects/opensource/performance-analysis/ogma-performance-analysis/node_modules/@ogma/styler/lib/styler.js:164:10
     35    0.3%    0.4%  LazyCompile: *writeGeneric node:internal/stream_base_commons:147:22
     27    0.2%    0.3%  RegExp: [^\n]$
      3    0.0%    0.0%  LazyCompile: *afterWrite node:internal/streams/writable:487:20
      1    0.0%    0.0%  Function: ^onwrite node:internal/streams/writable:425:17
      1    0.0%    0.0%  Function: ^normalizeString node:path:66:25
      1    0.0%    0.0%  Function: ^formatStream /home/h4ad/Projects/opensource/performance-analysis/ogma-performance-analysis/node_modules/@ogma/logger/src/logger/ogma.js:151:17
      1    0.0%    0.0%  Function: ^Socket._writeGeneric node:net:791:42
```

> [Source](https://github.com/jmcdo29/ogma/files/10368728/ogma-initial-solution.txt)

With the optimizations, it became:

```
Statistical profiling result from ogma-final-solution.log, (8995 ticks, 8 unaccounted, 0 excluded).

 [Shared libraries]:
   ticks  total  nonlib   name
   1792   19.9%          /home/h4ad/.asdf/installs/nodejs/16.16.0/bin/node
    604    6.7%          /usr/lib/x86_64-linux-gnu/libc-2.31.so
     53    0.6%          [vdso]

 [JavaScript]:
   ticks  total  nonlib   name
    124    1.4%    1.9%  LazyCompile: *formatStream /home/h4ad/Projects/opensource/performance-analysis/ogma-performance-analysis/node_modules/@ogma/logger/src/logger/ogma.js:152:17
     49    0.5%    0.7%  LazyCompile: *_write node:internal/streams/writable:283:16
     48    0.5%    0.7%  LazyCompile: *get magenta /home/h4ad/Projects/opensource/performance-analysis/ogma-performance-analysis/node_modules/@ogma/styler/lib/styler.js:123:16
     42    0.5%    0.6%  LazyCompile: *<anonymous> /home/h4ad/Projects/opensource/performance-analysis/ogma-performance-analysis/dist/ogma-log.stress.js:1:1
     41    0.5%    0.6%  LazyCompile: *writeGeneric node:internal/stream_base_commons:147:22
     38    0.4%    0.6%  LazyCompile: *get cyan /home/h4ad/Projects/opensource/performance-analysis/ogma-performance-analysis/node_modules/@ogma/styler/lib/styler.js:135:13
     24    0.3%    0.4%  RegExp: [^\n]$
      3    0.0%    0.0%  LazyCompile: *afterWrite node:internal/streams/writable:487:20
      1    0.0%    0.0%  Function: ^Socket._writeGeneric node:net:791:42
      1    0.0%    0.0%  Function: ^Socket._write node:net:828:35
```

> [Source](https://github.com/jmcdo29/ogma/files/10368737/ogma-final-solution.txt)

It didn't change too much, only the number of ticks and some functions.

When we compare the benchmarks, now we see some improvements, before:

```
'NestJS Logger x 7,214 ops/sec ±1.36% (86 runs sampled)',
'Ogma x 85,376 ops/sec ±4.57% (72 runs sampled)',
'Pino x 776,318 ops/sec ±4.46% (77 runs sampled)',
'console.log x 110,891 ops/sec ±2.17% (78 runs sampled)',
'process.stdout x 274,788 ops/sec ±3.16% (77 runs sampled)'
```

After:

```
'NestJS Logger x 7,300 ops/sec ±1.46% (82 runs sampled)',
'Ogma x 102,430 ops/sec ±3.98% (66 runs sampled)',
'Pino x 738,057 ops/sec ±6.79% (76 runs sampled)',
'console.log x 114,343 ops/sec ±2.44% (78 runs sampled)',
'process.stdout x 258,215 ops/sec ±2.32% (65 runs sampled)'
```

> [Benchmarks](https://gist.github.com/H4ad/38fc55fae31b4f0b77660094bdb17fbe)

Almost 20% of improvement, getting very close to `console.log`.

## Improvements that could introduce breaking change

All the improvements I made, I tried to not modify your API, so I only improve the existing code.

If you want to introduce breaking changes, in the future, you can refactor `styler.ts` to use methods instead `getters`, see the difference in performance:

```
'Getter x 67,877,439 ops/sec ±1.48% (89 runs sampled)',
'Method x 1,078,455,416 ops/sec ±0.48% (89 runs sampled)'
```

> This is a pure performance and is not a benchmark of your library.

Also, these lines:

https://github.com/jmcdo29/ogma/blob/a76f0805d755771e666540a7a2bf0f16ca8e676b/packages/logger/src/logger/ogma.ts#L123-L130

I don't know if you are familiar with the concept of [Hidden Classes](https://blog.bitsrc.io/secret-behind-javascript-performance-v8-hidden-classes-ba4d0ebfb89d) in V8, but in general, using `delete` operator is very heavy and is not recommended.

Also, you could initialize all properties of `json` object with `undefined` to indicate to V8 to reutilize the same Hidden Class across every option in the log, in the way that is now, the initial Hidden Class created for `json` will be mutated every time you add a new property.

Finally, your library will reach a maximum of performance using `process.stdout` eventually, so I recommend you to read the source code of `pino` to see what they are doing because their library is even faster than writing directly to the `stdout`.