@jmcdo29 2023-01-08
If CI fails and it's the `report-coverage` step then don't worry about it